Consolidate scanner utilities into shared backend
