+++ tools/scanner_diagnostics.py
@@ -13,7 +13,7 @@ import serial
 from serial.tools import list_ports
 
 from scanner_gui.controller import ScannerController
-from scanner_gui.scanner_utils import read_response, send_command, wait_for_data
+from utilities.scanner.backend import read_response, send_command, wait_for_data
 
 # Configure logging
 logging.basicConfig(
