+++ utilities/scanner/manager.py
@@ -9,7 +9,7 @@ import logging
 import serial
 
 from utilities.core.command_registry import build_command_table
-from utilities.core.scanner_utils import find_all_scanner_ports
+from utilities.scanner.backend import find_all_scanner_ports
 from utilities.io.timeout_utils import with_timeout
 from utilities.scanner.factory import get_scanner_adapter
